try to emulate with crateNumber ordering swapped
Details:
this is to test the hypothesis that one of the fibers is swapped on
one of the processors. Also add protecting '()' around the selection
cuts to avoid mistakes in the evaluation order of the selection
string, and print more info about the jets.

Note to self: there is a bug that shows up when processing multiple
files. The bug is due to the difference between `treeEntry` and
`chainEntry` of a TEntryList, but I wasn't able to fix it quickly (see
TEntryList doc and github.com/gerbaudo/root_utils/indexed_chain.py).